Continued to set up DB. Worked on User Schema
