+++ server/db/db.js
@@ -4,7 +4,7 @@
 // Import mongoose 
 import mongoose from 'mongoose';
 // Import Team Model
-import Team from './mongo/UserModel';
+import Team from './mongo/TeamModel';
 
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -20,7 +20,5 @@ const db = mongoose.connection;
 // Bind connection to error event (to get notification of connection errors)
 db.on('error', console.error.bind(console, 'MongoDB connection error:'));
 
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// SQL Section
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// Export db
+export default db;
