+++ server/server.js
@@ -2,6 +2,9 @@ const path = require('path');
 const express = require('express');
 const app = express();
 const port = 3000;
+const db = require('./db/db.js');
+const User = require('./db/mongo/UserModel.js');
+const Team = require('./db/mongo/TeamModel.js');
 
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
